enable syscalls/linux to support arm64
